Make the caching web proxy a service
This pull request moves the proxy system into a
full service that can be driven by config.  You can
externalize it.  Upstream-proxy continues to function as the
upstream for the proxy configured on the system.  The proxy can made
completely external as well.

You can not turn off the proxy.  It is required.