This adds delete support for NodeRoles.
The new rules for deleting things:

1. You can only delete noderoles that either:
   * Have no children
   * Have only children that are in PROPOSED and that have never run.
   * Have only children that are on the same node.

   These rules apply recursively down the noderole graph.

2. You can only delete deploymentroles that do not have noderoles
dependent on them.

3. Deleting a deployment deletes all the noderoles in it (using the
rules above, then the deploymentroles, then the phantom node for the
deployment, then it moves all the nodes to the deployment's parent.

All the new functionality has been exposed via the API (and the CLI),
bit not via the UI.  Deleting a thing via the CLI is accomplished using

    crowbar <objectclass> destroy <objectid>