Switch from delayed_jobs to que for running background tasks.
delayed_jobs was not playing nice with the transactional semantics we
rely on to ensure that only one job runs for a given node at a time, (it
was possible for several running tasks to be assigned to a single
running backend if the tasks weere all scheduled at the exact same tine
inside a transaction, which should not have been possible) and it was
also not preserving dedicated workers for HighPriorityWorker like we
wanted.

Instead of working on delayed_jobs to beat into shape, switch to que for
running background tasks instead.  Que is much simpler than
delayed_jobs, and it fully utilizes postgresql's advisory locking
semantics to ensure that the several tasks to one worker situation
cannot arise.

As an added bonus, que uses threads as backend workers instead of
processes, so it is lighter weight.