Refactor how dependent attrib poking works.
Now, all of the logic for figuring out when and how to poke dependent
noderoles based on what attribs they want is (mostly) contained in the
Attrib model, instead of being spread out across the node and noderole
classes.  Additionally, support was added for attribs to not poke
dependent noderoles if they should not.