Refactor network support to ressurect proper network role run behaviour.
Recent updates and refactors to the network handling code managed to
lose the property that any network noderole run knowns about the proper
configuration of all the networks bound to a node, and can change all
the configs as they see fit.  As a consequence, 2 important things were
lost:

1. We no longer automatically knew how to remove a network that we had
deliberatly removed from a node.
2. We no longer write proper network config files.

This pull request restores that behavior by adding a couple of custom
per-node attribs that keep track of the desired networking config, and
the networking config that we last placed on a node.