Add universal simple object matching to the Crowbar API.
Each controller grew a "match" POST route.  If you POST a JSON blob to
the route, it will return all the objects that match the attributes
passed in the blob.  The matching performed is simple equality

Despite the choice of verb, nothing is modified on the server side -- I
just needed a verb that expects a body, as the JSON can be arbitrarily
complex and thus not really passable in URL parameters.